UI框架逻辑重构

+++ Assets/Script/UI/UIConfig.cs
@@ -16,5 +16,5 @@ public class UIConfig
 
     public static readonly float ScreenHeightRatio = ResolutionHeight / Screen.height;
 
-    public static readonly float ScreenRatio = Screen.width / Screen.height;
-} 
+    public static readonly float ScreenRatio = Screen.width / (float)Screen.height;
+}
